add int exchange for special charartors in object name.

+++ KS3SDKIOS/KS3YunSDK/KS3YunSDK/KS3YunSDK/KSYS3/KSYModel/Object/KS3DownLoad.m
@@ -48,7 +48,7 @@
     
     const unsigned char *source = (const unsigned char *)[str UTF8String];
     
-    int sourceLen = strlen((const char *)source);
+    int sourceLen = (int)strlen((const char *)source);
     
     for (int i = 0; i < sourceLen; ++i) {
